NIFI-14277 Added credential scope in GCP PubSub and BigQuery processors
The credential scope is required when a private endpoint is used.

- also used ComponentLog in FlowFileResult in order to emit bulletin error logs
- also removed separate AbstractGCPubSubWithProxyProcessor which is no longer needed since PubSubLite processors have been removed
