+++ nifi-extension-bundles/nifi-gcp-bundle/nifi-gcp-processors/src/main/java/org/apache/nifi/processors/gcp/pubsub/PublishGCPubSub.java
@@ -106,7 +106,7 @@ import static org.apache.nifi.processors.gcp.pubsub.PubSubAttributes.TOPIC_NAME_
 })
 @SystemResourceConsideration(resource = SystemResource.MEMORY, description = "The entirety of the FlowFile's content "
         + "will be read into memory to be sent as a PubSub message.")
-public class PublishGCPubSub extends AbstractGCPubSubWithProxyProcessor {
+public class PublishGCPubSub extends AbstractGCPubSubProcessor {
     private static final List<String> REQUIRED_PERMISSIONS = Collections.singletonList("pubsub.topics.publish");
     private static final String TRANSIT_URI_FORMAT_STRING = "gcp://%s";
 
@@ -333,7 +333,7 @@ public class PublishGCPubSub extends AbstractGCPubSubWithProxyProcessor {
             if (flowFile.getSize() > maxMessageSize) {
                 final String message = String.format("FlowFile size %d exceeds MAX_MESSAGE_SIZE", flowFile.getSize());
                 failures.add(new IllegalArgumentException(message));
-                flowFileResults.add(new FlowFileResult(flowFile, futures, successes, failures));
+                flowFileResults.add(new FlowFileResult(flowFile, futures, successes, failures, getLogger()));
             } else {
                 baos.reset();
                 session.exportTo(flowFile, baos);
@@ -341,7 +341,7 @@ public class PublishGCPubSub extends AbstractGCPubSubWithProxyProcessor {
                 final ApiFuture<String> apiFuture = publishOneMessage(context, flowFile, baos.toByteArray());
                 futures.add(apiFuture);
                 addCallback(apiFuture, new TrackedApiFutureCallback(successes, failures), executor);
-                flowFileResults.add(new FlowFileResult(flowFile, futures, successes, failures));
+                flowFileResults.add(new FlowFileResult(flowFile, futures, successes, failures, getLogger()));
             }
         }
         finishBatch(session, stopWatch, flowFileResults);
@@ -391,7 +391,7 @@ public class PublishGCPubSub extends AbstractGCPubSubWithProxyProcessor {
                     futures.add(apiFuture);
                     addCallback(apiFuture, new TrackedApiFutureCallback(successes, failures), executor);
                 }
-                flowFileResults.add(new FlowFileResult(flowFile, futures, successes, failures));
+                flowFileResults.add(new FlowFileResult(flowFile, futures, successes, failures, getLogger()));
             }
         }
         finishBatch(session, stopWatch, flowFileResults);
